requirements.txt
streamlit
google-generativeai
pandas
matplotlib
seaborn
scikit-learn
kaggle